ai_image_detector/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
streamlit>=1.20.0
scikit-learn>=1.2.0
albumentations>=1.3.0
pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
